[Chore] #145- 코드 수정
