+++ Peekabook/.swiftlint.yml
@@ -7,6 +7,7 @@ disabled_rules:
     - line_length
     - function_parameter_count
     - function_body_length
+    - discouraged_direct_init
 
 opt_in_rules:
     - empty_count
