+++ Peekabook/Peekabook/Global/Utils/switchRootViewController.swift
@@ -1,0 +1,29 @@
+//
+//  switchRootViewController.swift
+//  Peekabook
+//
+//  Created by devxsby on 2023/01/24.
+//
+
+import UIKit
+
+extension UIViewController {
+    
+    func switchRootViewController(rootViewController: UIViewController, animated: Bool, completion: (() -> Void)?) {
+        guard let window = UIApplication.shared.windows.filter({ $0.isKeyWindow }).first else { return }
+        if animated {
+            UIView.transition(with: window, duration: 0.5, options: .transitionCrossDissolve, animations: {
+                let oldState: Bool = UIView.areAnimationsEnabled
+                UIView.setAnimationsEnabled(false)
+                window.rootViewController = rootViewController
+                UIView.setAnimationsEnabled(oldState)
+            }, completion: { (finished: Bool) -> Void in
+                if completion != nil {
+                    completion!()
+                }
+            })
+        } else {
+            window.rootViewController = rootViewController
+        }
+    }
+}
