+++ Peekabook/Peekabook/Presentation/AddBook/Views/BookInfoTableViewCell.swift
@@ -7,6 +7,7 @@
 
 import UIKit
 
+import Kingfisher
 import SnapKit
 import Then
 
@@ -20,7 +21,7 @@ class BookInfoTableViewCell: UITableViewCell {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         setUI()
         setLayout()
-        configButton()
+        config()
     }
     
     required init?(coder: NSCoder) {
@@ -32,20 +33,21 @@ class BookInfoTableViewCell: UITableViewCell {
     private let imgContainerView = UIView()
     private let bookImgView = UIImageView().then {
         $0.layer.masksToBounds = false
-        $0.layer.shadowOffset = CGSize(width: 1, height: 1)
-        $0.layer.shadowRadius = 4
-        $0.layer.shadowOpacity = 0.3
+        $0.layer.applyShadow(color: .black, alpha: 0.25, x: 1, y: 1, blur: 4, spread: 0)
         $0.contentMode = .scaleAspectFit
     }
+    
     private let labelContainerView = UIView()
     private let addContainerView = UIView()
     private let bookTitleLabel = UILabel().then {
-        $0.font = .h3
+        $0.numberOfLines = 2
+        $0.lineBreakMode = .byWordWrapping
+        $0.font = UIFont.font(.notoSansBold, ofSize: 15)
         $0.textColor = .peekaRed
     }
     
     private let authorLabel = UILabel().then {
-        $0.font = .h2
+        $0.font = .s1
         $0.textColor = .peekaRed
     }
     
@@ -55,9 +57,7 @@ class BookInfoTableViewCell: UITableViewCell {
         $0.text = "내 책장에 추가하기"
     }
     
-    private lazy var addButton = UIButton().then {
-        $0.addTarget(self, action: #selector(addButtonDidTap), for: .touchUpInside)
-    }
+    private let addButton = UIButton()
     
     override func awakeFromNib() {
         super.awakeFromNib()
@@ -78,13 +78,13 @@ extension BookInfoTableViewCell {
         contentView.backgroundColor = .white
         contentView.layer.borderWidth = 2
         contentView.layer.borderColor = UIColor.peekaRed.cgColor
-        
         imgContainerView.layer.borderWidth = 1
         imgContainerView.layer.borderColor = UIColor.peekaRed.cgColor
         
         labelContainerView.layer.borderWidth = 1
         labelContainerView.layer.borderColor = UIColor.peekaRed.cgColor
     }
+    
     private func setLayout() {
         self.backgroundColor = .clear
         contentView.addSubviews([
@@ -124,6 +124,7 @@ extension BookInfoTableViewCell {
         bookTitleLabel.snp.makeConstraints { make in
             make.top.equalToSuperview().offset(15)
             make.leading.equalToSuperview().offset(15)
+            make.width.equalTo(199)
         }
         
         authorLabel.snp.makeConstraints { make in
@@ -149,18 +150,15 @@ extension BookInfoTableViewCell {
         }
     }
     
-    private func configButton() {
+    private func config() {
         addButton.setImage(ImageLiterals.Icn.addBookMini, for: .normal)
     }
     
     func dataBind(model: BookInfoModel) {
         bookTitleLabel.text = model.title
         authorLabel.text = model.author
-        bookImgView.image = model.image
-    }
-    
-    @objc
-    private func addButtonDidTap() {
-        // dosomething
+        let url = URL(string: model.image)!
+        bookImgView.kf.indicatorType = .activity
+        bookImgView.kf.setImage(with: url)
     }
 }
