+++ Peekabook/Peekabook/Presentation/BookShelf/Views/CVC/BookShelfCVC.swift
@@ -79,9 +79,7 @@ extension BookShelfCVC {
     }
     
     func setData(model: Book) {
-        
         pickImageView.isHidden = model.pickIndex == 0 ? true : false
-        
         bookId = model.bookID
         bookImageView.kf.indicatorType = .activity
         bookImageView.kf.setImage(with: URL(string: model.book.bookImage))
