--- Peekabook/Peekabook/Presentation/Proposal/Views/.gitkeep.md
@@ -1,2 +1,0 @@
-#  <#Title#>
-
